extending the saved quantities in steps to include the REF module info + include the possibility of applying energy intercalibrations to REF module
