+++ cfg/moduleCharacterization_base.cfg
@@ -4,6 +4,7 @@ fileBaseName run
 runs runNumbers
 discCalibration /eos/cms/store/group/dpg_mtd/comm_mtd/TB/MTDTB_H8_Sep2025/TOFHIR/configs/confNumber/disc_calibration.tsv
 interCalibration interCalibPath
+interCalibration_ref interCalibPath_ref
 maxEntries -1
 step1FileName Lab5015_repo_path/plots/moduleCharacterization_step1_generalLabel.root
 usePedestals 0
